Created a separate file for Log In Page and methods for some widgets for clarity
